requirements.txt
google-generativeai
torch
torchvision
transformers
pillow
ultralytics
pytesseract
opencv-python
python-pptx
pypdf
pandas
openpyxl
